requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
billiard==4.2.0
black==23.11.0
celery==5.3.6
Cerberus==1.3.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==5.0.2
Django==5.0
django-allauth==0.59.0
django-appconf==1.0.6
django-autoslug==1.9.9
django-celery-email==3.0.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-environ==0.11.2
django-filter==23.5
django-phonenumber-field==7.2.0
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
flake8==6.1.0
flower==2.0.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
isort==5.13.0
kombu==5.3.4
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.0
pep517==0.13.1
phonenumbers==8.13.26
Pillow==10.1.0
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requirementslib==3.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomlkit==0.12.3
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
yarg==0.1.9
